\name{getPointColorsFromLatticeObject}
\alias{getPointColorsFromLatticeObject}
\title{Get Point Colors from Lattice Object}
\usage{
  getPointColorsFromLatticeObject(plot)
}
\arguments{
  \item{plot}{the lattice plot object.}
}
\description{
  Returns the colors used for points in a lattice plot
  object.
}

