\name{rearrangePanelXYPoints}
\alias{rearrangePanelXYPoints}
\title{Rearrange Points within Panel}
\usage{
  rearrangePanelXYPoints(plot)
}
\arguments{
  \item{plot}{the \code{lattice} plot object.}
}
\description{
  The x and y points within a panel are re-arranged so that
  their order matches the order \code{makeSVG} actually
  plots the points.
}

