\name{SVG.getLatticePointColors}
\alias{SVG.getLatticePointColors}
\title{Get and Convert Lattice Plot colors to RGB}
\usage{
  SVG.getLatticePointColors(plot)
}
\arguments{
  \item{plot}{the \code{lattice} plot object.}
}
\description{
  Gets the colors used for points in a lattice plot object,
  then converts them to the same format as expected in SVG
  attributes.
}

