\name{workingWithBoxPlots}
\alias{workingWithBoxPlots}
\title{Check if Current Plot Object is a Boxplot}
\usage{
  workingWithBoxPlots(plot)
}
\arguments{
  \item{plot}{the \code{lattice} plot object.}
}
\description{
  This function checks whether the current lattice plot is,
  or contains, boxplots. Works for single and multi-panel
  boxplots.
}

